requirements.txt
python-dotenv
ttkthemes
folium
tkinterweb
googlemaps
python-dotenv
pillow